backend/requirements.txt
Django==4.1.3
djangorestframework==3.14.0
Pillow==9.3.0
psycopg2-binary==2.9.5
redis==4.3.4
sorl-thumbnail==12.9.0